feat: add build and release support
This commit adds a makefile to handle all manner of package maintenance. It also adds a circle.yml to allow releases to occur automatically through circle-ci using gh-release.

We also gitignore the build directory to avoid accidentally committing binaries to the repository.

Finally, it fleshes out the readme to provide users with better usage context.