Correct the whatis header for NAME section
manpages have this special section for `NAME`
In markdown this will look like this (tabs added for readability):

  # NAME
  <name> - <description>

So we have a heading with a value of `NAME` followed by some text
Where <name> is the command or list of commands with comma separated values.
The command names must not contain spaces.
The `-` before the description must be escaped with `\`.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>